heptacert/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-multipart==0.0.12

SQLAlchemy==2.0.36
asyncpg==0.29.0
alembic==1.13.3
psycopg2-binary==2.9.9

pydantic==2.9.2
pydantic-settings==2.6.1

bcrypt==3.2.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

Pillow==10.4.0
qrcode==7.4.2
pandas==2.2.3
openpyxl==3.1.5
email-validator==2.2.0
itsdangerous==2.2.0
aiosmtplib==3.0.2
httpx==0.27.2
slowapi==0.1.9
pyotp==2.9.0
pyhanko==0.21.0
pyhanko-certvalidator==0.26.3
APScheduler==3.10.4
cryptography==43.0.3
geoip2==4.8.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
aiosqlite==0.20.0